crypto_bot_requirements.txt
# Crypto Trading Bot Requirements
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiohttp>=3.9.0
pyjwt>=2.8.0
python-dotenv>=1.0.0
numpy>=1.26.0
websockets>=12.0
python-multipart>=0.0.9
